backend/requirements.txt
Flask
flask-cors
rembg
Pillow
numpy
scikit-learn